# services/dbt_web/openapi/dbt_web.openapi.yaml
openapi: 3.0.3
info:
  title: dbt-web API
  version: 0.1.0
servers:
  - url: http://localhost:8010
paths:
  /api/v1/runs/{target}:
    post:
      summary: Trigger dbt run by target
      parameters:
        - in: path
          name: target
          required: true
          schema:
            type: string
            enum: [staging, vault, marts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunJobRequest"
      responses:
        "200":
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunJobResponse"
  /api/v1/runs/{run_id}:
    get:
      summary: Get run status
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunStatusResponse"
  /api/v1/runs/{run_id}/logs:
    get:
      summary: Get run logs
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLogsResponse"
  /api/v1/runs/{run_id}/artifacts/{name}:
    get:
      summary: Get run artifact
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
            enum: [manifest.json, catalog.json, run_results.json, graph.js]
      responses:
        "200":
          description: Artifact payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactEnvelope"
  /api/v1/events/ingestion_completed:
    post:
      summary: Airflow ingestion event webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventPayload"
      responses:
        "200":
          description: Accepted
  /api/v1/events/datavault_completed:
    post:
      summary: Airflow datavault event webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventPayload"
      responses:
        "200":
          description: Accepted
  /api/v1/events/marts_completed:
    post:
      summary: Airflow marts event webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventPayload"
      responses:
        "200":
          description: Accepted
components:
  schemas:
    RunJobRequest:
      type: object
      properties:
        selectors:
          type: array
          items:
            type: string
        vars:
          type: object
          additionalProperties:
            type: string
        full_refresh:
          type: boolean
        defer:
          type: boolean
        fail_on_test_failure:
          type: boolean
    RunJobResponse:
      type: object
      properties:
        run_id: { type: string }
        status: { type: string }
        submitted_at: { type: string, format: date-time }
        target: { type: string }
    RunStatusResponse:
      type: object
      properties:
        run_id: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        duration_sec: { type: number }
        job_name: { type: string }
        artifacts:
          type: array
          items: { type: string }
    RunLogsResponse:
      type: object
      properties:
        run_id: { type: string }
        lines:
          type: array
          items: { type: string }
        truncated: { type: boolean }
        updated_at: { type: string, format: date-time }
    ArtifactEnvelope:
      type: object
      properties:
        run_id: { type: string }
        name: { type: string }
        etag: { type: string }
        size: { type: integer }
        content_type: { type: string }
        cached: { type: boolean }
        content: {}
    EventPayload:
      type: object
      required: [event_id, dag_id, event_ts]
      properties:
        event_id: { type: string }
        dag_id: { type: string }
        run_id: { type: string }
        event_ts: { type: string, format: date-time }
        target_layer: { type: string }
        upstream_run_ref: { type: string }